requirements.txt
discord
aiosqlite
python-dateutil
requests
aiohttp
Pillow
discord-components